requirements.txt
absl-py==2.4.0
alembic==1.18.4
astunparse==1.6.3
attrs==26.1.0
banal==1.0.6
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
contourpy==1.3.2
cv==1.0.0
cvzone==1.6.1
cycler==0.12.1
dataset==1.6.2
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.78.0
h5py==3.14.0
idna==3.11
jax==0.6.2
jaxlib==0.6.2
joblib==1.5.3
kaggle==1.7.4.5
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.11
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.31.1
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.5
SQLAlchemy==1.4.54
tensorflow==2.21.0
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
wrapt==2.1.2
